refactor: jsのモジュールファイルの名前を変更
