+++ app/javascript/packs/modules/task_form_requestable.js
@@ -1,4 +1,4 @@
-import requestByConfiguredAxios from './configured_axios';
+import requestByConfiguredAxios from './request_by_configured_axios';
 import toastr from 'toastr';
 
 export default {
