+++ app/javascript/packs/tasks/create_task_form_vue.js
@@ -1,5 +1,5 @@
 import Vue from 'vue/dist/vue.esm.js';
-import taskFormRequestable from '../modules/task_form_mixin';
+import taskFormRequestable from '../modules/task_form_requestable';
 
 window.task_form = new Vue({
   el: '#task_form',
